Implemented NoteDBHelper as Singleton; Fixed NoteDAOImplSQLite getAllSavedNotes method
New instances of NoteDAOImplSQLite were originally creating new instances of NoteDBHelper,
and thus of the DB itself. NoteDAOImplSQLite was called in both MainActivity and EditNoteActivity.
Hence multiple instances of the SQLite db were left open, warning was thrown.
NoteDBHelper was implemented as a singleton, now only its getInstance method is called.
Prevents any other instances of the DB from being generated in the application.

getAllSavedNotes method was reading only the first record and writing that data to every
notePreview retrieved by the noteGridAdapter for display in the MainActivity. Now works well.

Ongoing Issue: the multi-select delete method in MainActivity fails.