Moved Code for Multi-Delete feature from MainActivity into NoteGridAdapter
Originally had code for creating and ArrayList of the Notes to be deleted in MainActivity,
in the OnClick callback method for the Action Bar Options Menu.
Since this method takes positions of checked items in the Adapter and returns a list of the notes at those positions,
this should properly be a noteGridAdapter member method.

Created a public member field for noteGridAdapter in which to store this ArrayList, and an updateCheckedNoteList method
to be called whenever the list of currently checked items needs to be stored.
MainActivity now only calls DAO methods directly.