UI issue in MainActivity Resolved; Photo Resolution consistent, faster
The flagged issues regarding the overlapping notes, the mixed up images,
and missing images were resolved by rewriting the custom NoteGridAdapter,
and the XML layout for the note previews that show in MainActivity.

XML Layouts: now have a fixed height relative to the display which is
fetched programmatically.

Dynamically changing Layouts: Layout paramters of the views now change
dynamically depending on whether the note has an image or not. Note text
is shown in place of the image when image not available.

Titles: of notes now have ellipses if they run out of bounds. Now bold, navy.

Photo Resolution: has now been fixed for each activity relative to the
size of the display. this fixes a note size relative to the particular screen size
ensuring a consistent experience across different devices.

Issues: only remaining issue is the lag when multiple notes are selected.